+++ utils.go
@@ -39,7 +39,7 @@ func Index[T comparable](slice []T, toFind T) int {
 	return -1
 }
 
-func posAlreadySeen(slice []Node, toFind [][]uint8) Node {
+func posAlreadySeen(slice []Node, toFind [][]int) Node {
 	for _, v := range slice {
 		if isEqual(v.world, toFind) {
 			return v
@@ -111,7 +111,7 @@ func matrixToString2(matrix [][]int) string {
 	return results
 }
 
-func matrixToStringSelector(matrix [][]uint8, worker int, seenNodeMap int) (key string, queueIndex int, seenNodeIndex int) {
+func matrixToStringSelector(matrix [][]int, worker int, seenNodeMap int) (key string, queueIndex int, seenNodeIndex int) {
 	if len(matrix) < 10 {
 		return matrixToString(matrix, worker, seenNodeMap)
 	} else {
@@ -120,7 +120,7 @@ func matrixToStringSelector(matrix [][]uint8, worker int, seenNodeMap int) (key 
 }
 
 //3224
-func matrixToString(matrix [][]uint8, worker int, seenNodeMap int) (key string, queueIndex int, seenNodeIndex int) {
+func matrixToString(matrix [][]int, worker int, seenNodeMap int) (key string, queueIndex int, seenNodeIndex int) {
 
 	results := make([]byte, len(matrix)*len(matrix)*4)
 	size := len(matrix)
@@ -129,8 +129,8 @@ func matrixToString(matrix [][]uint8, worker int, seenNodeMap int) (key string, 
 	for i := 0; i < size; i++ {
 
 		for j := 0; j < size; j++ {
-			queueIndex += int(matrix[i][j]) * i * j
-			seenNodeIndex += int(matrix[i][j]) * i * j
+			queueIndex += matrix[i][j] * i * j
+			seenNodeIndex += matrix[i][j] * i * j
 			results[spot] = byte(matrix[i][j] / 10)
 			results[spot+1] = byte(matrix[i][j] % 10)
 			results[spot+2] = '.'
@@ -142,7 +142,7 @@ func matrixToString(matrix [][]uint8, worker int, seenNodeMap int) (key string, 
 	return string(results), queueIndex, seenNodeIndex
 }
 
-func matrixToStringNoOpti(matrix [][]uint8, worker int, seenNodeMap int) (key string, queueIndex int, seenNodeIndex int) {
+func matrixToStringNoOpti(matrix [][]int, worker int, seenNodeMap int) (key string, queueIndex int, seenNodeIndex int) {
 
 	results := ""
 	size := len(matrix)
@@ -150,9 +150,9 @@ func matrixToStringNoOpti(matrix [][]uint8, worker int, seenNodeMap int) (key st
 	for i := 0; i < size; i++ {
 
 		for j := 0; j < size; j++ {
-			queueIndex += int(matrix[i][j]) * i * j
-			seenNodeIndex += int(matrix[i][j]) * i * j
-			results += strconv.Itoa(int(matrix[i][j])) + "."
+			queueIndex += matrix[i][j] * i * j
+			seenNodeIndex += matrix[i][j] * i * j
+			results += strconv.Itoa(matrix[i][j]) + "."
 
 		}
 	}
